Day-5/Travel Assistant/requirements.txt
streamlit==1.28.0
requests==2.31.0
langchain==0.1.0
langchain-community==0.0.20
langchain-google-genai==0.0.11
langchain-core==0.1.0
duckduckgo-search==3.9.6
python-dotenv==1.0.0